# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "矿山自动化调度系统 API",
    "description": "面向矿山运输调度场景的数字化管理 API，覆盖运单、排队、装车、称重、看板、告警、调度算法等完整业务链路。",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "开发环境" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/api/v1/drivers": {
      "get": {
        "summary": "司机列表（分页）",
        "tags": ["司机管理"],
        "parameters": [
          { "name": "keyword", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["idle", "working", "offline"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ]
      },
      "post": {
        "summary": "创建司机",
        "tags": ["司机管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "phone", "license_plate", "vehicle_type", "capacity_ton"],
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "license_plate": { "type": "string" },
                  "vehicle_type": { "type": "string", "enum": ["dump_truck", "trailer", "other"] },
                  "capacity_ton": { "type": "number" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/drivers/import": {
      "post": {
        "summary": "批量导入司机",
        "tags": ["司机管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["source", "drivers"],
                "properties": {
                  "source": { "type": "string" },
                  "drivers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["name", "phone", "license_plate", "vehicle_type", "capacity_ton"]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/loading/waybills/{waybill_id}/start": {
      "post": {
        "summary": "开始装车",
        "tags": ["装车管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["operator_id"],
                "properties": {
                  "operator_id": { "type": "string", "format": "uuid" },
                  "loader_name": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loading/waybills/{waybill_id}/finish": {
      "post": {
        "summary": "结束装车",
        "tags": ["装车管理"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["operator_id"],
                "properties": {
                  "operator_id": { "type": "string", "format": "uuid" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}